pyproject.toml
[tool.poetry]
name = "canadian-accessibility-advisor"
version = "0.1.0"
description = "Canadian accessibility law advisor"
authors = ["yukaty"]
readme = "README.md"
packages = [{ include = "advisor", from = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^0.3"
langgraph = "^0.2"
langchain-openai = "^0.3"
langchain-community = "^0.3"
faiss-cpu = "^1.9"
streamlit = "^1.40"
python-dotenv = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
